+++ Observer/CheckoutOnePageControllerSuccessActionEventObserver.php
@@ -7,6 +7,7 @@ use Magento\Framework\Event\Observer;
 use Magento\Framework\App\ObjectManager;
 use Magento\Sales\Model\OrderRepository;
 use Gracious\Interconnect\Observer\ObserverAbstract;
+use Gracious\Interconnect\Generic\Behaviours\SendsOrder;
 use Gracious\Interconnect\Http\Request\Client as InterconnectClient;
 use Gracious\Interconnect\Http\Request\Data\Order\Factory as OrderDataFactory;
 
@@ -16,6 +17,8 @@ use Gracious\Interconnect\Http\Request\Data\Order\Factory as OrderDataFactory;
  */
 class CheckoutOnePageControllerSuccessActionEventObserver extends ObserverAbstract
 {
+    use SendsOrder;
+
     /**
      * {@inheritdoc}
      */
@@ -35,30 +38,8 @@ class CheckoutOnePageControllerSuccessActionEventObserver extends ObserverAbstra
         }
 
         $orderId = $orderIds[0];
-        $order = $orderRepository = ObjectManager::getInstance()->create(OrderRepository::class)->get($orderId);
-        $orderDataFactory = new OrderDataFactory();
+        /* @var $order Order */ $order = $orderRepository = ObjectManager::getInstance()->create(OrderRepository::class)->get($orderId);
 
-        try {
-            $requestData = $orderDataFactory->setupData($order);
-        }catch (Throwable $exception) {
-//            $this->logger->error('Failed to prepare the order data. *** MESSAGE ***:  '.$exception->getMessage().',  *** TRACE ***: '.$exception->getTraceAsString());
-            $this->logger->error('Failed to prepare the order data. *** MESSAGE ***:  '.$exception->getMessage());
-
-            return;
-        }
-
-        $this->logger->debug('Order data: ' . json_encode($requestData));
-        
-        // Using try/catch because we don't want this to interfere with critical logic (for example: crash the checkout so that orders can not be placed)
-        try {
-            $this->client->sendData($requestData, InterconnectClient::ENDPOINT_ORDER);
-        }catch (Throwable $exception) {
-//            $this->logger->error('Failed to send order. *** MESSAGE ***: '.$exception->getMessage().', *** TRACE ***: '.$exception->getTraceAsString());
-            $this->logger->error('Failed to send order. *** MESSAGE ***: '.$exception->getMessage());
-
-            return;
-        }
-
-        $this->logger->info(__METHOD__.' :: Order sent to Interconnect ('.$order->getId().')');
+        $this->sendOrder($order, $this->logger, $this->client);
     }
 }