+++ Reporting/Logger.php
@@ -3,6 +3,11 @@ namespace Gracious\Interconnect\Reporting;
 
 use \Monolog\Logger as Base;
 
+/**
+ * Class Logger
+ * @package Gracious\Interconnect\Reporting
+ * Custom logger (yes this class is empty but that's the way it works with Magento so don't delete it)
+ */
 class Logger extends Base
 {
 }