+++ Reporting/Logger/Handler.php
@@ -4,6 +4,11 @@ namespace Gracious\Interconnect\Reporting\Logger;
 use \Monolog\Logger;
 use \Magento\Framework\Logger\Handler\Base;
 
+/**
+ * Class Handler
+ * @package Gracious\Interconnect\Reporting\Logger
+ * Custom logger handler so we have our own log for this module.
+ */
 class Handler extends Base
 {
     /**